chore: rename helper modules with behavior-revealing names
Replace ambiguous *-helper.ts and *-helpers.ts module names with names that describe what they do:

Core replication:
- onedrive-replication-helpers.ts → onedrive-replication-result.ts (converts results to status records)
- sharepoint-replication-helpers.ts → sharepoint-replication-result.ts
- outlook-replication-helpers.ts → outlook-manifest-handler.ts (handles manifest operations)
- rehydration-dek-helper.ts → dek-rehydration-validator.ts (validates DEK state)

M365 Graph:
- graph-error-helpers.ts → graph-request-error-handler.ts (classifies/retries Graph errors)
  - Rename corresponding test: graph-error-helpers.test.ts → graph-request-error-handler.test.ts

OneDrive:
- graph-onedrive-download-helpers.ts → graph-onedrive-download-executor.ts (handles large file downloads)

SharePoint:
- graph-sharepoint-download-helpers.ts → graph-sharepoint-download-executor.ts

S3:
- s3-error-helpers.ts → s3-error-classifier.ts (classifies S3 errors)

Update 28 import paths across packages and tests.

Per CLAUDE.md: 'Do not use ambiguous *.helper.ts names in services. Use behavior-revealing names.' This change is enforceable in future lint rules.

Fixes #156 (partial - helper naming only; JSDoc and core re-exports in separate PRs)
